Ensure markup options for type are always available
When ActiveRecord/ActiveModel determines the types for attributes,
part of the code which does so instantiates a new instance of the type
class. This mainly seems to happen when there are multiple attributes
of the same type on a record.

The upshot though is that we can't use an instance variable of the
type class to store markup options and pass them on to the typecast
markup string.

In order to store the markup options on the class, we need to use a
different class for each combination of markup options, which means
generating new classes for each possible combination of markup
options. This can be done on demand, using a stable key generation
mechanism.