Ensure the specs actually run on Travis
For some reason, the `spec` task that appears in Travis is not
actually the one defined in the Rakefile. When we run it with
`--trace` on Travis, we can see this:

    $ bundle exec rake --trace
    ** Invoke load_app (first_time)
    ** Execute load_app
    ** Invoke default (first_time)
    ** Invoke spec (first_time, not_needed)
    ** Execute default

That `not_needed` is a sign that it's not actually going to be run,
and I'm not sure why.

In the short term, we can just use `rspec` directly.